remove service interfaces
